5_pinecone_rag_with_mcp_tools/requirements.txt
# Google ADK - Agent Development Kit
google-adk==1.14.1
google-adk[a2a]

# Environment Variables
python-dotenv==1.1.0

# Pinecone - Vector Database with Integrated Embedding
# Note: Package renamed from 'pinecone-client' to 'pinecone' in v5.1.0
pinecone>=5.0.0

# Note: ไม่ต้องใช้ openai เพราะใช้ Pinecone integrated embedding!
